Remove unnecessary `exact` attributes
Exact is useful in certain situations. For example, if you have two routes ('/' and '/users'), and the '/' route is listed first, then the router will match any path beginning with '/'. When you use the exact property, the router will check to make sure your path is EXACTLY '/' instead of a false positive like '/users'.

Same goes with if you have a route for '/users', but you have another route for '/users/all_user' . You'd want to use the `exact` property on the '/user' route.