Fix the value of "Set Type", change values on select
This is what I get for commiting before getting ready for work. The
value was just 't'. Now the value is the `workout` const from the hook,
which is '' by default. `setWorkout` changes it via `handleChange` in
the field's `onChange`. Right? I know we are still learning hook basics
but I feel like I get it.