Return user to previous page with back arrow in `CreateRecord FC`
Uses the history api to go back to the previous route. Instead of
being in a button, a back arrow icon has a an `onClick` to go back.
An actual `<button>`/`<Button>` didn't look right and seemed like
overkill, but it may be better practice to reconsider it for semantics.