Start EditExercise component skeleton and validation
A big commit, but I needed to sketch out alot for it to fit together. 

Basically, the rendered JSX is my attempt to copy the component from the old app... AND much of it will probably be divided into smaller components (pending).

There is also a simple implementation of validation in the input for the exercise name.