Alternative edit-exercise component
This is a rough sketch of the overarching component. Much what is rendered will be moved to smaller components. The dynamic forms (warm-up/working sets) will almost certainly be their own components. We just need to figure out how the information flows.

Also, check out the basic implementation of validation.